+++ apps/ungdomsytelse-deltaker/src/apps/søknad/SøknadApp.tsx
@@ -12,7 +12,10 @@ import { AppRoutes } from '../../utils/AppRoutes';
 import { SøknadProvider } from './context/SøknadContext';
 import { useBarn } from './hooks/api/useBarn';
 import { useKontonummer } from './hooks/api/useKontonummer';
+import { HarKontonummerEnum } from './steg/oppsummering/oppsummeringUtils';
 import SøknadRouter from './SøknadRouter';
+import { KontonummerOppslagInfo } from './types';
+import { formaterKontonummer } from './utils/formaterKontonummer';
 
 const SøknadApp = () => {
     const { søker, deltakelsePeriode } = useDeltakerContext();
@@ -34,17 +37,35 @@ const SøknadApp = () => {
         return <UngLoadingPage />;
     }
 
-    if (barn.isError || kontonummer.isError) {
-        if (barn.isError) {
-            const { type, message, context } = barn.error;
-            logApiError(ApiErrorKey.barn, { type, message, context });
-        }
-        if (kontonummer.isError) {
-            const { type, message, context } = kontonummer.error;
-            logApiError(ApiErrorKey.kontonummer, { type, message, context });
-        }
+    if (barn.isError) {
+        const { context, message, type } = barn.error;
+        logApiError(ApiErrorKey.barn, { error: { context, message, type } });
         return <HentDeltakerErrorPage error={text('søknadApp.loading.error')} />;
     }
+
+    // if (kontonummer.isError && isApiAxiosError(kontonummer.error) && kontonummer.error.originalError.status !== 503) {
+    //     const { context, message, type } = kontonummer.error;
+    //     logApiError(ApiErrorKey.kontonummer, { error: { context, message, type } });
+    //     return <HentDeltakerErrorPage error={text('søknadApp.loading.error')} />;
+    // }
+
+    const getKontonummerInfo = (): KontonummerOppslagInfo => {
+        if (kontonummer.error) {
+            return {
+                harKontonummer: HarKontonummerEnum.UVISST,
+            };
+        }
+
+        return kontonummer.data?.harKontonummer && kontonummer.data.kontonummer
+            ? {
+                  harKontonummer: HarKontonummerEnum.JA,
+                  kontonummerFraRegister: kontonummer.data.kontonummer,
+                  formatertKontonummer: formaterKontonummer(kontonummer.data.kontonummer),
+              }
+            : {
+                  harKontonummer: HarKontonummerEnum.NEI,
+              };
+    };
 
     const søknadOppgave = deltakelsePeriode.oppgaver.find((o) => o.oppgavetype === Oppgavetype.SØK_YTELSE);
 
@@ -58,7 +79,7 @@ const SøknadApp = () => {
                 søknadOppgave={søknadOppgave}
                 søker={søker}
                 deltakelsePeriode={deltakelsePeriode}
-                kontonummer={kontonummer.data?.harKontonummer ? kontonummer.data.kontonummer : undefined}
+                kontonummerInfo={getKontonummerInfo()}
                 barn={barn.data || []}>
                 <SøknadRouter />
             </SøknadProvider>