+++ apps/ungdomsytelse-deltaker/src/apps/søknad/hooks/api/useBarn.ts
@@ -11,7 +11,8 @@ export const useBarn = (enabled = true) => {
         queryKey: søknadQueryKeys.barn,
         queryFn: () => fetchBarn(),
         enabled,
-        staleTime: 1000 * 60 * 5, // 5 minutter
+        staleTime: 1000 * 60 * 20, // 20 minutter
         retry: 1,
+        refetchOnWindowFocus: false,
     });
 };