+++ apps/ungdomsytelse-deltaker/src/apps/søknad/steg/oppsummering/oppsummeringUtils.ts
@@ -1,31 +1,73 @@
-import { ungdomsytelse } from '@navikt/k9-brukerdialog-prosessering-api';
+import { KontonummerInfo, ungdomsytelse } from '@navikt/k9-brukerdialog-prosessering-api';
 import { YesOrNo } from '@navikt/sif-common-formik-ds';
 import { dateToISODate } from '@navikt/sif-common-utils';
 import { Spørsmål, SøknadSvar } from '../../types';
-import { formaterKontonummer } from '../../utils/formaterKontonummer';
 
 const isYesOrNoAnswered = (answer?: YesOrNo) => {
     return answer === YesOrNo.YES || answer === YesOrNo.NO;
 };
 
-export const buildSøknadFromSvar = (
-    deltakelseId: string,
-    oppgaveReferanse: string,
-    svar: SøknadSvar,
-    søkerNorskIdent: string,
-    startdato: Date,
-    kontonummerFraRegister?: string,
-): Omit<ungdomsytelse.Ungdomsytelsesøknad, 'harBekreftetOpplysninger'> | undefined => {
-    const harKontonummer = !!kontonummerFraRegister;
-    if (
-        svar[Spørsmål.FORSTÅR_PLIKTER] !== true ||
-        !isYesOrNoAnswered(svar[Spørsmål.BARN]) ||
-        (harKontonummer && !isYesOrNoAnswered(svar[Spørsmål.KONTONUMMER]))
-    ) {
+export type HarKontonummerValues = Pick<KontonummerInfo, 'harKontonummer'>;
+
+export enum HarKontonummerEnum {
+    JA = 'JA',
+    NEI = 'NEI',
+    UVISST = 'UVISST',
+}
+
+export type SøknadApiData = Omit<ungdomsytelse.Ungdomsytelsesøknad, 'harBekreftetOpplysninger' | 'kontonummerErRiktig'>;
+
+export const getKontonummerApiInfo = (
+    kontonummerInfo: KontonummerInfo,
+    kontonummerErRiktigSvar?: YesOrNo,
+): KontonummerInfo | undefined => {
+    switch (kontonummerInfo.harKontonummer) {
+        case 'UVISST':
+            return {
+                harKontonummer: 'UVISST',
+            };
+        case 'JA':
+            if (!isYesOrNoAnswered(kontonummerErRiktigSvar)) {
+                return undefined;
+            }
+            return {
+                harKontonummer: 'JA',
+                kontonummerFraRegister: kontonummerInfo.kontonummerFraRegister,
+                kontonummerErRiktig: kontonummerErRiktigSvar === YesOrNo.YES,
+            };
+        case 'NEI':
+            return {
+                harKontonummer: 'NEI',
+            };
+    }
+};
+
+export const buildSøknadFromSvar = ({
+    deltakelseId,
+    oppgaveReferanse,
+    svar,
+    søkerNorskIdent,
+    startdato,
+    kontonummerInfo,
+}: {
+    deltakelseId: string;
+    oppgaveReferanse: string;
+    svar: SøknadSvar;
+    søkerNorskIdent: string;
+    startdato: Date;
+    kontonummerInfo: KontonummerInfo;
+}): SøknadApiData | undefined => {
+    if (svar[Spørsmål.FORSTÅR_PLIKTER] !== true || !isYesOrNoAnswered(svar[Spørsmål.BARN])) {
         return undefined;
     }
-
     const harForståttRettigheterOgPlikter = svar[Spørsmål.FORSTÅR_PLIKTER] === true;
+    const kontonummerApiInfo = getKontonummerApiInfo(kontonummerInfo, svar[Spørsmål.KONTONUMMER]);
+
+    if (!kontonummerApiInfo) {
+        // eslint-disable-next-line no-console
+        console.error('Kontonummer info is missing or invalid');
+        return undefined;
+    }
 
     return {
         deltakelseId,
@@ -34,8 +76,7 @@ export const buildSøknadFromSvar = (
         startdato: dateToISODate(startdato),
         harForståttRettigheterOgPlikter,
         barnErRiktig: svar[Spørsmål.BARN] === YesOrNo.YES,
-        kontonummerErRiktig: harKontonummer ? svar[Spørsmål.KONTONUMMER] === YesOrNo.YES : undefined,
+        kontonummerInfo: kontonummerApiInfo,
         søkerNorskIdent,
-        kontonummerFraRegister: kontonummerFraRegister ? formaterKontonummer(kontonummerFraRegister) : undefined,
     };
 };