+++ packages/k9-brukerdialog-prosessering-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/k9-brukerdialog-prosessering-api
 
+## 0.1.0
+
+### Minor Changes
+
+- 6acb88a: Oppdaterte genererte typer
+
 ## 0.0.23
 
 ### Patch Changes